+++ common/cosmetics/hearts.ts
@@ -1,7 +1,7 @@
 import CantTouchThis from '../achievements/cant-touch-this'
 import DeckedOut from '../achievements/decked-out'
 import SUStainable from '../achievements/sustainable'
-import Channeling from '../achievements/targeted-lightning'
+import Channeling from '../achievements/channeling'
 import Win from '../achievements/wins'
 import {Heart} from './types'
